+++ src/components/auth/AuthForm.tsx
@@ -1,7 +1,7 @@
-import { useErrorMessage } from '@/hooks/auth/useErrorMessage';
-import { useInput } from '@/hooks/auth/useInput';
 import { useSetAuth } from '@/hooks/auth/useSetAuth';
 import { useValid } from '@/hooks/auth/useValid';
+import { useErrorMessage } from '@/hooks/service/auth/useErrorMessage';
+import { useInput } from '@/hooks/service/auth/useInput';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { FaCheck } from 'react-icons/fa6';
